Move Validation class to ErrorHandling package
