+++ src/main/java/com/github/aksc/Grammar/Symbol.java
@@ -2,6 +2,7 @@ package com.github.aksc.Grammar;
 
 import com.github.aksc.DerivationSystem;
 import com.github.aksc.ErrorHandling.BadLanguageException;
+import com.github.aksc.ErrorHandling.Validation;
 import com.github.aksc.MetaData.*;
 import com.google.gson.annotations.SerializedName;
